Move PublicRole.behavior to behavior directory.
