+++ test/access/roles/CapperRole.test.js
@@ -1,4 +1,4 @@
-const { shouldBehaveLikePublicRole } = require('../../access/roles/PublicRole.behavior');
+const { shouldBehaveLikePublicRole } = require('../../behavior/access/roles/PublicRole.behavior');
 const CapperRoleMock = artifacts.require('CapperRoleMock');
 
 contract('CapperRole', function ([_, capper, otherCapper, ...otherAccounts]) {
