+++ test/lifecycle/Pausable.test.js
@@ -1,5 +1,5 @@
 const { expectEvent, shouldFail } = require('openzeppelin-test-helpers');
-const { shouldBehaveLikePublicRole } = require('../access/roles/PublicRole.behavior');
+const { shouldBehaveLikePublicRole } = require('../behavior/access/roles/PublicRole.behavior');
 
 const PausableMock = artifacts.require('PausableMock');
